+++ src/coords_manager.cu
@@ -119,4 +119,73 @@ CoordsManager::getUnionInOutMapsGPU(vector<py::object> py_in_coords_keys,
   return make_pair(ref(d_in_maps[map_key]), ref(d_out_maps[map_key]));
 }
 
+/*
+ * Given tensor_stride_src and tensor_stride_dst, find the respective coord_maps
+ * and return the indices of the coord_map_ind in coord_map_dst
+ */
+vector<vector<at::Tensor>> CoordsManager::getKernelMapGPU(
+    vector<int> tensor_strides, vector<int> strides, vector<int> kernel_sizes,
+    vector<int> dilations, int region_type, at::Tensor offsets,
+    py::object py_in_coords_key, py::object py_out_coords_key,
+    bool is_transpose, bool is_pool) {
+  // WARNING: This function will not work properly with custon region types.
+  ASSERT(region_type != 2,
+         "Currently, it does not support the custom region type.");
+  const InOutMapKey map_key = getMapHashKey(
+      tensor_strides, strides, kernel_sizes, dilations, region_type,
+      py_in_coords_key, py_out_coords_key, is_transpose, is_pool);
+
+  const auto &in_out = getInOutMapsGPU(
+      tensor_strides, strides, kernel_sizes, dilations, region_type, offsets,
+      py_in_coords_key, py_out_coords_key, false);
+
+  const pInOutMaps<int> &in_maps = in_out.first;
+  const pInOutMaps<int> &out_maps = in_out.second;
+
+  int all_volume = 0, kernel_volume = in_maps.size();
+  for (int k = 0; k < kernel_volume; k++)
+    all_volume += in_maps[k].size();
+
+  // CUDA_CHECK(cudaGetDevice(&device_id));
+  torch::TensorOptions options =
+      torch::TensorOptions()
+          .dtype(torch::kInt32)
+          // .device(torch::kCUDA)
+          .device(torch::kCUDA, gpu_memory_manager.device_id)
+          .requires_grad(false);
+
+  cudaStream_t stream = at::cuda::getCurrentCUDAStream();
+
+  vector<at::Tensor> in_tensors, out_tensors;
+
+  for (int k = 0; k < kernel_volume; k++) {
+    auto curr_volume = in_maps[k].size();
+    if (curr_volume <= 0)
+      continue;
+
+    at::Tensor in_kernel_map =
+        torch::empty({(long)curr_volume}, options).contiguous();
+    at::Tensor out_kernel_map =
+        torch::empty({(long)curr_volume}, options).contiguous();
+
+    // Wait until memory is allocated
+    CUDA_CHECK(cudaStreamSynchronize(stream));
+
+    int *p_in_kernel_map = in_kernel_map.data<int>();
+    int *p_out_kernel_map = out_kernel_map.data<int>();
+
+    CUDA_CHECK(cudaMemcpy(p_in_kernel_map, in_maps[k].data(),
+                          curr_volume * sizeof(int),
+                          cudaMemcpyDeviceToDevice));
+    CUDA_CHECK(cudaMemcpy(p_out_kernel_map, out_maps[k].data(),
+                          curr_volume * sizeof(int),
+                          cudaMemcpyDeviceToDevice));
+
+    in_tensors.push_back(move(in_kernel_map));
+    out_tensors.push_back(move(out_kernel_map));
+  }
+
+  return {in_tensors, out_tensors};
+}
+
 } // end namespace minkowski