+++ tests/kernel_map.py
@@ -23,12 +23,54 @@
 # of the code.
 import unittest
 
+import torch
+
 from MinkowskiEngine import SparseTensor, MinkowskiConvolution
 
 from tests.common import data_loader
 
 
 class TestKernelMap(unittest.TestCase):
+
+    def test_kernelmap_gpu(self):
+        print(f"{self.__class__.__name__}: test_kernelmap_gpu")
+        if not torch.cuda.is_available():
+            return
+
+        in_channels, out_channels, D = 2, 3, 2
+        coords, feats, labels = data_loader(in_channels)
+        feats = feats.double()
+        feats.requires_grad_()
+        input = SparseTensor(feats, coords=coords)
+        cm = input.coords_man
+        ikey = cm._get_coords_key(1)
+        print('Input coords: ')
+        cm.print_diagnostics(ikey)
+
+        print('Convolution: ')
+
+        # Initialize context
+        conv = MinkowskiConvolution(
+            in_channels,
+            out_channels,
+            kernel_size=3,
+            stride=2,
+            has_bias=True,
+            dimension=D).double()
+        output = conv(input)
+
+        iC = input.C.numpy()
+        oC = output.C.numpy()
+        print(iC)
+        print(oC)
+        in_maps, out_maps = output.coords_man.get_kernel_map(1, 2, stride=2, kernel_size=3, on_gpu=True)
+        kernel_index = 0
+        for in_map, out_map in zip(in_maps, out_maps):
+          for i, o in zip(in_map, out_map):
+            print(kernel_index, iC[i], '->', oC[o])
+          kernel_index += 1
+        self.assertTrue(sum(len(in_map) for in_map in in_maps) == 26)
+
 
     def test_kernelmap(self):
         print(f"{self.__class__.__name__}: test_kernelmap")